tests: test branch/tag/bookmark display in test-hgweb-filelog.t
Displaying branches, tags and bookmarks is an obviously important feature of
hgweb and should be tested a bit more than not at all, so let's add a branch, a
tag and a bookmark to the test.

With this change it's evident that the default style (paper) doesn't show
bookmarks in filelog. Future patch will fix this.