revset: make internal _list() expression remove duplicated revisions
This allows us to optimize chained 'or' operations to _list() expression.

Unlike _intlist() or _hexlist(), it's difficult to remove duplicates by the
caller of _list() because different symbols can point to the same revision.
If the caller knows all symbols are unique, that probably means revisions or
nodes are known, therefore, _intlist() or _hexlist() should be used instead.
So, it makes sense to check duplicates by _list() function.

'%ls' is no longer used in core, this won't cause performance regression.