test: rename 'test-push-r.t' to 'test-push.t'
We do not have a simple test for 'hg push' but we have multiple tiny tests for
various aspect of it. We'll unify them into a single file, and we start with
'test-push-r.t'. The code is unchanged but we renamed the repository used to
avoid collision with other tests we'll import in coming changesets.

Test timing for the record:

start   end     cuser   csys    real      Test
  1.850   2.640   0.650   0.090   0.790   test-push-validation.t
  2.640   3.520   0.760   0.090   0.880   test-push-hook-lock.t
  0.000   1.850   1.560   0.210   1.850   test-push-r.t